Clean up some issues and fix import issues.
