+++ tests/unit/static/golden/test_otel_instrument__test_instrument_decorator.csv
