+++ tests/unit/test_otel_instrument.py
@@ -2,17 +2,20 @@
 Tests for OTEL instrument decorator.
 """
 
-from typing import Any, Dict
 from unittest import main
 
 import pandas as pd
 import sqlalchemy as sa
 from trulens.apps.custom import TruCustomApp
+from trulens.core.schema.event import EventRecordType
 from trulens.core.session import TruSession
 from trulens.experimental.otel_tracing.core.init import init
 from trulens.experimental.otel_tracing.core.instrument import instrument
 
 from tests.test import TruTestCase
+from tests.util.df_comparison import (
+    compare_dfs_accounting_for_ids_and_timestamps,
+)
 
 
 class _TestApp:
@@ -55,9 +58,6 @@ class _TestApp:
 
 
 class TestOtelInstrument(TruTestCase):
-    def setUp(self):
-        pass
-
     @staticmethod
     def _get_events() -> pd.DataFrame:
         tru_session = TruSession()
@@ -68,8 +68,14 @@ class TestOtelInstrument(TruTestCase):
 
     @staticmethod
     def _convert_column_types(df: pd.DataFrame):
+        # Writing to CSV and the reading back causes some type issues so we
+        # hackily convert things here.
         df["event_id"] = df["event_id"].apply(str)
-        df["record_type"] = df["record_type"].apply(lambda x: eval(x))
+        df["record_type"] = df["record_type"].apply(
+            lambda x: EventRecordType(x[len("EventRecordType.") :])
+            if x.startswith("EventRecordType.")
+            else EventRecordType(x)
+        )
         df["start_timestamp"] = df["start_timestamp"].apply(pd.Timestamp)
         df["timestamp"] = df["timestamp"].apply(pd.Timestamp)
         for json_column in [
@@ -80,103 +86,7 @@ class TestOtelInstrument(TruTestCase):
         ]:
             df[json_column] = df[json_column].apply(lambda x: eval(x))
 
-    def _compare_dfs_accounting_for_ids_and_timestamps(
-        self, expected: pd.DataFrame, actual: pd.DataFrame
-    ):
-        """
-        Compare two Dataframes are equal, accounting for ids and timestamps.
-        That is:
-        1. The ids between the two Dataframes may be different, but they have
-           to be consistent. That is, if one Dataframe reuses an id in two
-           places, then the other must as well.
-        2. The timestamps between the two Dataframes may be different, but
-           they have to be in the same order.
-
-        Args:
-            expected: expected results
-            actual: actual results
-        """
-        id_mapping: Dict[str, str] = {}
-        timestamp_mapping: Dict[pd.Timestamp, pd.Timestamp] = {}
-        self.assertEqual(len(expected), len(actual))
-        self.assertListEqual(list(expected.columns), list(actual.columns))
-        for i in range(len(expected)):
-            for col in expected.columns:
-                self._compare_entity(
-                    expected.iloc[i][col],
-                    actual.iloc[i][col],
-                    id_mapping,
-                    timestamp_mapping,
-                    is_id=col.endswith("_id"),
-                    locator=f"df.iloc[{i}][{col}]",
-                )
-        # Ensure that the id mapping is a bijection.
-        self.assertEqual(
-            len(set(id_mapping.values())),
-            len(id_mapping),
-            "Ids are not a bijection!",
-        )
-        # Ensure that the timestamp mapping is monotonic.
-        prev_value = None
-        for curr in sorted(timestamp_mapping.keys()):
-            if prev_value is not None:
-                self.assertLess(
-                    prev_value,
-                    timestamp_mapping[curr],
-                    "Timestamps are not in the same order!",
-                )
-            prev_value = timestamp_mapping[curr]
-
-    def _compare_entity(
-        self,
-        expected: Any,
-        actual: Any,
-        id_mapping: Dict[str, str],
-        timestamp_mapping: Dict[pd.Timestamp, pd.Timestamp],
-        is_id: bool,
-        locator: str,
-    ):
-        self.assertEqual(
-            type(expected), type(actual), f"Types of {locator} do not match!"
-        )
-        if is_id:
-            self.assertEqual(
-                type(expected), str, f"Type of id {locator} is not a string!"
-            )
-            if expected not in id_mapping:
-                id_mapping[expected] = actual
-            self.assertEqual(
-                id_mapping[expected],
-                actual,
-                f"Ids of {locator} are not consistent!",
-            )
-        elif isinstance(expected, dict):
-            self.assertEqual(
-                expected.keys(),
-                actual.keys(),
-                f"Keys of {locator} do not match!",
-            )
-            for k in expected.keys():
-                self._compare_entity(
-                    expected[k],
-                    actual[k],
-                    id_mapping,
-                    timestamp_mapping,
-                    is_id=k.endswith("_id"),
-                    locator=f"{locator}[k]",
-                )
-        elif isinstance(expected, pd.Timestamp):
-            if expected not in timestamp_mapping:
-                timestamp_mapping[expected] = actual
-            self.assertEqual(
-                timestamp_mapping[expected],
-                actual,
-                f"Timestamps of {locator} are not consistent!",
-            )
-        else:
-            self.assertEqual(expected, actual, f"{locator} does not match!")
-
-    def test_deterministic_app_id(self):
+    def test_instrument_decorator(self):
         # Set up.
         tru_session = TruSession()
         tru_session.experimental_enable_feature("otel_tracing")
@@ -190,13 +100,13 @@ class TestOtelInstrument(TruTestCase):
         with custom_app:
             test_app.respond_to_query("throw")
         # Compare results to expected.
-        GOLDEN_FILENAME = "tests/unit/static/golden/test_otel_instrument__test_deterministic_app_id.csv"
+        GOLDEN_FILENAME = "tests/unit/static/golden/test_otel_instrument__test_instrument_decorator.csv"
         actual = self._get_events()
         self.assertEqual(len(actual), 8)
         self.write_golden(GOLDEN_FILENAME, actual)
         expected = self.load_golden(GOLDEN_FILENAME)
         self._convert_column_types(expected)
-        self._compare_dfs_accounting_for_ids_and_timestamps(expected, actual)
+        compare_dfs_accounting_for_ids_and_timestamps(self, expected, actual)
 
 
 if __name__ == "__main__":
